# openapi.json
{
  "components": {
    "schemas": {
      "BacktestRequest": {
        "properties": {
          "include_holdout": {
            "default": true,
            "title": "Include Holdout",
            "type": "boolean"
          },
          "pattern_name": {
            "title": "Pattern Name",
            "type": "string"
          },
          "period_days": {
            "default": 730,
            "maximum": 3650.0,
            "minimum": 60.0,
            "title": "Period Days",
            "type": "integer"
          },
          "stop_loss": {
            "default": 0.02,
            "exclusiveMinimum": 0.0,
            "maximum": 0.5,
            "title": "Stop Loss",
            "type": "number"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "take_profit": {
            "default": 0.04,
            "exclusiveMinimum": 0.0,
            "maximum": 1.0,
            "title": "Take Profit",
            "type": "number"
          }
        },
        "required": [
          "symbol",
          "pattern_name"
        ],
        "title": "BacktestRequest",
        "type": "object"
      },
      "BacktestResponse": {
        "additionalProperties": false,
        "properties": {
          "equity_curve": {
            "items": {
              "$ref": "#/components/schemas/EquityPoint"
            },
            "title": "Equity Curve",
            "type": "array"
          },
          "holdout": {
            "$ref": "#/components/schemas/HoldoutInfo"
          },
          "is_valid": {
            "title": "Is Valid",
            "type": "boolean"
          },
          "metrics": {
            "additionalProperties": true,
            "title": "Metrics",
            "type": "object"
          },
          "metrics_display": {
            "additionalProperties": true,
            "title": "Metrics Display",
            "type": "object"
          },
          "pattern": {
            "title": "Pattern",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "trades": {
            "items": {
              "$ref": "#/components/schemas/TradeOut"
            },
            "title": "Trades",
            "type": "array"
          },
          "validation_status": {
            "default": "unvalidated",
            "title": "Validation Status",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "pattern",
          "metrics",
          "metrics_display",
          "equity_curve",
          "trades",
          "is_valid",
          "holdout"
        ],
        "title": "BacktestResponse",
        "type": "object"
      },
...
      "EquityPoint": {
        "additionalProperties": false,
        "properties": {
          "date": {
            "title": "Date",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "date",
          "value"
        ],
        "title": "EquityPoint",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "Body of every non-2xx response: ``{\"error\": {\"code\", \"message\", \"details\", \"request_id\"}}``.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HoldoutInfo": {
        "additionalProperties": false,
        "properties": {
          "bars": {
            "title": "Bars",
            "type": "integer"
          },
          "note": {
            "title": "Note",
            "type": "string"
          },
          "share": {
            "title": "Share",
            "type": "number"
          },
          "start": {
            "title": "Start",
            "type": "string"
          }
        },
        "required": [
          "start",
          "bars",
          "share",
          "note"
        ],
        "title": "HoldoutInfo",
        "type": "object"
      },
      "LatestMLResult": {
        "additionalProperties": false,
        "properties": {
          "age_hours": {
            "title": "Age Hours",
            "type": "number"
          },
          "generated_at": {
            "title": "Generated At",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "payload": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Payload",
            "type": "array"
          },
          "stale": {
            "title": "Stale",
            "type": "boolean"
          }
        },
        "required": [
          "kind",
          "generated_at",
          "age_hours",
          "stale",
          "payload"
        ],
        "title": "LatestMLResult",
        "type": "object"
      },
      "LatestPairsResult": {
        "additionalProperties": false,
        "properties": {
          "age_hours": {
            "title": "Age Hours",
            "type": "number"
          },
          "generated_at": {
            "title": "Generated At",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "payload": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Payload",
            "type": "array"
          },
          "stale": {
            "title": "Stale",
            "type": "boolean"
          }
        },
        "required": [
          "kind",
          "generated_at",
          "age_hours",
          "stale",
          "payload"
        ],
        "title": "LatestPairsResult",
        "type": "object"
      },
...
      "LatestTechnicalResult": {
        "additionalProperties": false,
        "description": "Latest nightly technical scan: the validated / OOS-positive candidates (a superset of ScanSignal).",
        "properties": {
          "age_hours": {
            "title": "Age Hours",
            "type": "number"
          },
          "generated_at": {
            "title": "Generated At",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "payload": {
            "items": {
              "$ref": "#/components/schemas/TechnicalCandidate"
            },
            "title": "Payload",
            "type": "array"
          },
          "stale": {
            "title": "Stale",
            "type": "boolean"
          }
        },
        "required": [
          "kind",
          "generated_at",
          "age_hours",
          "stale",
          "payload"
        ],
        "title": "LatestTechnicalResult",
        "type": "object"
      },
...
      "RegimeRequest": {
        "properties": {
          "period_days": {
            "default": 730,
            "maximum": 3650.0,
            "minimum": 60.0,
            "title": "Period Days",
            "type": "integer"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol"
        ],
        "title": "RegimeRequest",
        "type": "object"
      },
...
      "StressReport": {
        "additionalProperties": true,
        "description": "full_stress_test's report, passed through as one dict (its existing keys are the contract).",
        "title": "StressReport",
        "type": "object"
      },
      "StressRequest": {
        "properties": {
          "pattern_name": {
            "title": "Pattern Name",
            "type": "string"
          },
          "period_days": {
            "default": 2920,
            "maximum": 3650.0,
            "minimum": 60.0,
            "title": "Period Days",
            "type": "integer"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "pattern_name"
        ],
        "title": "StressRequest",
        "type": "object"
      },
...
      "TradeOut": {
        "additionalProperties": false,
        "properties": {
          "bars_held": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bars Held"
          },
          "direction": {
            "title": "Direction",
            "type": "string"
          },
          "entry_date": {
            "title": "Entry Date",
            "type": "string"
          },
          "entry_price": {
            "title": "Entry Price",
            "type": "number"
          },
          "exit_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exit Date"
          },
          "exit_price": {
            "title": "Exit Price",
            "type": "number"
          },
          "exit_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exit Reason"
          },
          "in_holdout": {
            "default": false,
            "title": "In Holdout",
            "type": "boolean"
          },
          "pnl_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pnl Pct"
          }
        },
        "required": [
          "entry_date",
          "direction",
          "entry_price",
          "exit_price"
        ],
        "title": "TradeOut",
        "type": "object"
      },
...
      "WalkForwardRequest": {
        "properties": {
          "n_splits": {
            "default": 4,
            "maximum": 12.0,
            "minimum": 2.0,
            "title": "N Splits",
            "type": "integer"
          },
          "pattern_name": {
            "title": "Pattern Name",
            "type": "string"
          },
          "period_days": {
            "default": 730,
            "maximum": 3650.0,
            "minimum": 60.0,
            "title": "Period Days",
            "type": "integer"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "pattern_name"
        ],
        "title": "WalkForwardRequest",
        "type": "object"
      },
      "WalkForwardResponse": {
        "additionalProperties": false,
        "properties": {
          "fold_scheme": {
            "title": "Fold Scheme",
            "type": "string"
          },
          "folds": {
            "items": {
              "$ref": "#/components/schemas/WalkForwardFold"
            },
            "title": "Folds",
            "type": "array"
          },
          "n_folds": {
            "title": "N Folds",
            "type": "integer"
          },
          "pattern": {
            "title": "Pattern",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "pattern",
          "n_folds",
          "folds",
          "fold_scheme"
        ],
        "title": "WalkForwardResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Trading Bot v2 API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/backtest/run": {
      "post": {
        "description": "Run a backtest on a symbol + pattern combination (execution convention: config.EXECUTION_MODE).",
        "operationId": "run_backtest_api_backtest_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthorized (API_TOKEN set, token missing or wrong)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unknown_symbol / no_data / not_found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation_error / unknown_pattern / insufficient_data"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "internal_error / contract_violation"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "data_quality"
          }
        },
        "summary": "Run Backtest",
        "tags": [
          "backtest"
        ]
      }
    },
    "/api/backtest/walk-forward": {
      "post": {
        "description": "Run walk-forward validation.",
        "operationId": "run_walk_forward_api_backtest_walk_forward_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WalkForwardRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalkForwardResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthorized (API_TOKEN set, token missing or wrong)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unknown_symbol / no_data / not_found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation_error / unknown_pattern / insufficient_data"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "busy: another heavy request is running"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "internal_error / contract_violation"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "data_quality"
          }
        },
        "summary": "Run Walk Forward",
        "tags": [
          "backtest"
        ]
      }
    },
...
    "/api/results/ml/latest": {
      "get": {
        "description": "Latest nightly ML scan.",
        "operationId": "latest_ml_api_results_ml_latest_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LatestMLResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthorized (API_TOKEN set, token missing or wrong)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unknown_symbol / no_data / not_found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation_error / unknown_pattern / insufficient_data"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "internal_error / contract_violation"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "data_quality"
          }
        },
        "summary": "Latest Ml",
        "tags": [
          "results"
        ]
      }
    },
    "/api/results/pairs/latest": {
      "get": {
        "description": "Latest nightly pairs scan.",
        "operationId": "latest_pairs_api_results_pairs_latest_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LatestPairsResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthorized (API_TOKEN set, token missing or wrong)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unknown_symbol / no_data / not_found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation_error / unknown_pattern / insufficient_data"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "internal_error / contract_violation"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "data_quality"
          }
        },
        "summary": "Latest Pairs",
        "tags": [
          "results"
        ]
      }
    },
    "/api/results/technical/latest": {
      "get": {
        "description": "Latest nightly technical scan (validated / OOS-positive candidates): payload, generated_at, stale flag.",
        "operationId": "latest_technical_api_results_technical_latest_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LatestTechnicalResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthorized (API_TOKEN set, token missing or wrong)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unknown_symbol / no_data / not_found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation_error / unknown_pattern / insufficient_data"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "internal_error / contract_violation"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "data_quality"
          }
        },
        "summary": "Latest Technical",
        "tags": [
          "results"
        ]
      }
    },
    "/api/stress/full": {
      "post": {
        "description": "Run the complete stress test suite.",
        "operationId": "run_stress_test_api_stress_full_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StressRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StressReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthorized (API_TOKEN set, token missing or wrong)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unknown_symbol / no_data / not_found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation_error / unknown_pattern / insufficient_data"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "busy: another heavy request is running"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "internal_error / contract_violation"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "data_quality"
          }
        },
        "summary": "Run Stress Test",
        "tags": [
          "stress"
        ]
      }
    },
    "/api/stress/regimes": {
      "post": {
        "description": "Detect market regimes for a symbol.",
        "operationId": "get_regimes_api_stress_regimes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegimeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Regimes",
        "tags": [
          "stress"
        ]
      }
    },
    "/api/stress/sensitivity": {
      "post": {
        "description": "Run parameter sensitivity analysis (bars before HOLDOUT_START only).",
        "operationId": "get_sensitivity_api_stress_sensitivity_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StressRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthorized (API_TOKEN set, token missing or wrong)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unknown_symbol / no_data / not_found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation_error / unknown_pattern / insufficient_data"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "busy: another heavy request is running"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "internal_error / contract_violation"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "data_quality"
          }
        },
        "summary": "Get Sensitivity",
        "tags": [
          "stress"
        ]
      }
    }
  }
}